test: Cars 단위 테스트 클래스 생성 및 테스트 코드 이동
